Simplify interface to the user.
The option to allow the user to batch DrawBox()
and DrawText() calls together was removed. Now,
every call to either of those functions issues
a draw-call to the renderer. Actually, this
did not result in worse performance. We can
later optimize this to batch multiple
calls to those functions together if we really
draw *a lot* text and 2d-shapes. Right now,
I realized it just makes both the usage-code
and the underlying OpenGL rendercode more
complicated with no substantial benefit and
potentially more bugs. We can look at this later
(see comments).