fix/chore: Fix header-only-lib include. Start door entity parsing.
It is important to have header-only-libs included at the top
of the file *before* all other includes. This makes sure
the preprocessor doesn't just include the declaration if
another include includes that header-only-lib. This would
'shadow' the include that has the IMPLEMENTATION #define
before it. Then, the linker cannot find the definitions
and fails to build.