refactor: Make renderer available via Getter function
Reason: We need to be able to issue render Commands from
anywhere in the program not just in game.cpp! We would
have to pass the render instance to all of the other
places and that is just not very nice.
It could potentially get a problem when splitting
up the game into its own DLL but if that is the
case it is also not hard to fix: We just pass
the instance pointer of the renderer to the game
again and then make sure that the game itself
exposes that instance via a Getter function
to the other places.