fix: Dont' create a new world-batch for each texture type.
We put all triangles that share the same texture into
their dedicated batch when loading the map. Thus,
draw calls could be reduced to the number of textures.
This is the right idea but it is an anit-pattern to
allocate new memory on the GPU all the time when
the data is exactly the same (triangles). Also
we just don't know how much memory on the GPU is
needed for each texture-type. Also all these allocations
are pretty slow, *especially* when tested on
integrated AMD GPUs!!! Which is kinda weird...
So now we got rid of that approach and insted put *all*
of the triangles in *one* world-batch. Then, we
create self-defined draw commands that hold the
offset into this one world-batch. We use
a map to know for which texture-id (GPU handle)
what draw commands have to be used. This could
also be optimized further by not using a map
at all which causes a ton of cache misses
(especially STLs very dumb implementation)
and just put the texture id into the draw
commands themselves. For now, this already
is a very good improvement!