feat: Delegate input to specific entities.
Introduce an interface, IInputReceiver, that
an entity has to implement if it wants to
receive high-level input commands (such
as "fire", "forward", etc.). Also, a
input-gouvernor is in place, the
CInputDelegate singleton class. This
one can be configured to route high-
level input commands to specific
entities that implement the IInputReceiver.