fix: Crash on unknown command or command without args
The program crashed in PrintfImpl due to not
enough space allocated for the formatted buffer.
vsnprintf does not return the number of bytes
needed for all of the args (only the fmt characters
are counted without replacing %s with the actual
args value). I believe the args would have to be
dealt with and copied seperately.