fix: set next waypoint on path init.
The global path (created in CWorld) will get
its next waypoint assigned to whatever comes
after the current waypoint. This is random
because in a cycling path there is no 'first'.
However, an entity gets assigned a path because
that entity has a target to a path's waypoint
then the next waypoint is whatever the
entity's initial (set in TrenchBroom) points to.