fix: This finally seems to work!
Collision detection now behaves the
same, no matter the framerate. However,
at this point there are quite a few
hacks and so many updates happening
it is hard to understand the code. I
have it all in my head now and I need
to clean this up before I forget.
So don't even bother looking at it
at this point. This is just
a 'good state' in case I mess things
up again (and I will).