fix/wip: Start integrating enemies into new collision code.
- We need to update the ellipsoid collider per entity
so we don't use the const version of GetEllipsoidCollider().
- Refactor the collision code so it takes *all* entities
into account.
- The entity manger has a new update function that is called
before the collision code is being run. This is to setup
the desired velocity vector for the entities. The collision
code needs this desired velocity vector (m_Velocity of
BaseGameEntity) to do its thing. The Update() needs to be refactored
into another name. As J.Gregory says in Game Engine
Architecture, multiple update steps are needed. There is not
really a way around this!
- The current enemy code is only partially working for this
commit.