feat: Better 'in air' vs 'on ground' check for enemy/player.
A sphere is traced down (0.0f, 0.0f, -1.0f)*MIN_DISTANCE_TO_GROUND
to check if entity is in 'flying' or 'on ground' state. This
very accurate as eg. stair climbing was previously considered
as 'in air' as the velocity vector went upwards. This also
made it necessary to play with magic values until a
certain slop would not be considered as 'in air'. The current
approach, however is *very* expensive as the sphere is traced
against the *entire* world again. A geometry subdivision is
very much in need...