fix: Add original soloud.h and rename it to soloud.h.bak.
We need to modify soloud.h as the original #includes
windows.h. This has the implication that symbols
like 'min' and 'max' as being used in glm::min, glm::max
are 'shadowed' as windows.h #defines min/max (thanks
microsoft). The thing is upon saving my (Michael) editor
reformats the code according to .clang-format and the
.clang-format-ignore (included in this commit) did not
work.