refactor: additional file for drawing frustum (and other things).
In order not to blow up the irender interface with even
more virtual functions I introduced a new file called 'r_render'
that takes uses existing render functions defined in the
render-interface. For now, only code for drawing a frustum
is in there. But other functions probably could be taken
out of the render interface and moved to r_draw.